LayoutTests/http/tests/websocket/tests/hybi/binary-type-expected.txt

CONSOLE MESSAGE: line 600: 'Blob' is not a valid value for binaryType; binaryType remains unchanged.
CONSOLE MESSAGE: line 600: 'ArrayBuffer' is not a valid value for binaryType; binaryType remains unchanged.
CONSOLE MESSAGE: line 600: '' is not a valid value for binaryType; binaryType remains unchanged.
Test WebSocket.binaryType attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws.binaryType is "blob"
PASS ws.binaryType is "arraybuffer"
PASS ws.binaryType is "blob"
PASS ws.binaryType = 'Blob' threw exception Error: SyntaxError: DOM Exception 12.
PASS ws.binaryType is "blob"
PASS ws.binaryType = 'ArrayBuffer' threw exception Error: SyntaxError: DOM Exception 12.
PASS ws.binaryType is "blob"
PASS ws.binaryType = '' threw exception Error: SyntaxError: DOM Exception 12.
PASS ws.binaryType is "blob"
PASS successfullyParsed is true

TEST COMPLETE
